do not redirect to /dev/null
So we can check & determine whether the error is desired or is caused by
stricter compiler check. Later case should be fixed in source file.

In case of choose.sh, redirect stdout to stderr because script's stdout will be
redirect to a source file which will be used in later compilation