show full build command and do not redirect to /dev/null
Show full build command so we can check the build option.

Remove redirect to /dev/null, so we can check & determine whether the
error is desired or is caused by stricter compiler check. Later case
should be fixed in source file.

In case of choose.sh, redirect stdout to stderr because script's stdout will be
redirect to a source file which will be used in later compilation